+++ packages/ai-kit/src/components/ai-kit-service/locales/en.ts
@@ -6,5 +6,6 @@ export default flat({
   aiProviderSubscription: 'Subscribe to AI Service',
   aiProviderLocalAIKit: 'Local AIKit component',
   selectMonth: 'Select Month',
+  subscribeAITip: 'Subscribe to AI Kit now to unlock the power of AI!',
   subscribeAIService: 'Subscribe AI Service',
 });